Get endianness

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -12,5 +12,5 @@ int get_endianness(void)
 	unsigned int value = 1;
 	char *ptr = (char *)&value;
 
-	return (int)(*ptr);
+	return (*ptr);
 }
